feat: Add ContributionGameSVG class to create Lizard and ContributionMap groups for simulation
A new class, `ContributionGameSVG`, has been added to the project. This class
is responsible for creating and managing two main groups of SVG elements:
a `Lizard` group and a `ContributionMap` group.

The purpose of this class is to simulate the behavior of the procedural
lizard interacting with the contribution map. The simulation will depict
the lizard "eating" each contribution point on the map, providing a
unique and engaging visual representation of a user's activity. This
merges two previously distinct functionalities to create a new,
interactive demonstration.